Små rettelser: koordinater til party bliver hentet ned ved edit og add. Party sætter nu koordinater korrekt (som double)

+++ FOGmain/src/dk/vinael/domain/Party.java
@@ -255,8 +255,8 @@ public class Party implements Serializable {
 			setMaxAge(obj.getInt("max_age"));
 			setShowPhotos(obj.getInt("show_photos"));
 			setShowWall(obj.getInt("show_wall"));
-			setLat(obj.getInt("lat"));
-			setLon(obj.getInt("lon"));
+			setLat(obj.getDouble("lat"));
+			setLon(obj.getDouble("lon"));
 		} catch (JSONException e) {
 			e.printStackTrace();
 		}
